requirements.txt
camelot-py[cv]
geopy
folium
flask
flask-bootstrap4
flask-socketio
